Reconfigure MVC layout

+++ controllers/coasterController.js
@@ -1,16 +1,14 @@
-const db = require('../db/coasterQueries');
-const { getRiderIdFromName, addCoasterToRiderById } = require('../db/riderQueries');
+const db = require('../models/coasterModels');
+const { getRiderIdFromName, addCoasterToRiderById } = require('../models/riderModels');
 const { validationResult, matchedData } = require('express-validator');
+const success = require('../utils/success');
+const failure = require('../utils/failure');
 
 // Retrieves all coasters from database and displays it on coasters main page
 async function getAllCoasters(req, res, next) {
     try {
         const coasters = await db.getAllCoasters();
-        res.render('allCoasters', {
-            title: 'All Coasters',
-            coasters: coasters,
-        });
-
+        return success(res, 200, 'allCoasters', 'All Coasters', coasters);
     } catch (err) {
         next(err);
     };
@@ -19,10 +17,7 @@ async function getAllCoasters(req, res, next) {
 // Retrieves the add new coaster form
 async function getNewCoasterForm(req, res, next) {
     try {
-        res.render('newCoasterForm', {
-            title: 'Add New Coaster',
-        });
-
+        return success(res, 200, 'newCoasterForm', 'Add New Coaster')
     } catch (err) {
         next(err);
     };
@@ -33,22 +28,21 @@ async function postNewCoaster(req, res, next) {
     try {
         // Gets results from coaster form validation in middleware/validateCoaster.js
         const errors = validationResult(req);
-
         // If there are any errors, redisplay the form page with errors at the top
         if (!errors.isEmpty()) {
-            return res.status(400).render('newCoasterForm', {
-                title: 'Add New Coaster',
-                errors: errors.array()
-            });
+            return failure(res, 400, 'newCoasterForm', 'Add New Coaster', errors.array())
         };
-
         // Receive sanitized and verified information and use it to create new coaster entry in database
         const { coasterName, coasterInversions, coasterSpeed, coasterHeight, coasterLength, parkName } = matchedData(req);
-        await db.postNewCoaster(coasterName, coasterInversions, coasterSpeed, coasterHeight, coasterLength, parkName);
+        const parkId = await db.getOrCreateParkId(parkName);
+        await db.postNewCoaster(coasterName, coasterInversions, coasterSpeed, coasterHeight, coasterLength, parkName, parkId);
         // Returns to main coaster page after submitting
-        res.redirect('/coaster');
-
+        return success(res, 201, '/coaster');
     } catch (err) {
+        if (err.code === '23505') {
+            return failure(res, 409, 'errors', 'Error 409 - Coaster already exists')
+        };
+
         next(err);
     };
 };
@@ -58,23 +52,14 @@ async function getUpdateCoasterForm(req, res, next) {
     try {
         // Receives validated ID from middleware/validateId.js
         const id = req.validatedId;
-
         // Finds matching coaster in the database by ID
         const coaster = await db.findCoasterById(id);
-
         // If ID doesn't match any coasters, redirects to the error page
         if (!coaster) {
-            return res.status(401).render('errors', {
-                title: 'Error 401 - Coaster not found',
-                message: 'Error 401 - Coaster not found in database',
-            });
+            return failure(res, 401, 'errors', 'Error 401 - Coaster not found');
         };
-
         // Once coaster has been found, renders the update form with the coaster details
-        res.render('updateCoasterForm', {
-            title: 'Update Coaster',
-            coaster: coaster,
-        });
+        return success(res, 200, 'updateCoasterForm', `Update ${coaster.name}`, coaster);
 
     } catch (err) {
         next(err);
@@ -83,41 +68,33 @@ async function getUpdateCoasterForm(req, res, next) {
 
 // Adds updated coaster data to the database based on the id from the previous form retrieval, 
 // adds a new park to the parks table if the park name field doesn't find a matching name
-async function postUpdatedCoaster(req, res, next) {
+async function putUpdatedCoaster(req, res, next) {
     try {
         // Gets results from coaster form validation in middleware/validateCoaster.js
         const errors = validationResult(req);
         // Receives validated ID from middleware/validateId.js
         const id = req.validatedId;
-
+        // Finds matching coaster in the database by ID
+        const coaster = await db.findCoasterById(id);
+        // If ID doesn't match any coasters, redirects to the error page
+        if (!coaster) {
+            return failure(res, 401, 'errors', 'Error 401 - Coaster not found');
+        };
         // If there are any errors, redisplay the form page with errors at the top
         if (!errors.isEmpty()) {
-            // Finds matching coaster in the database by ID
-            const coaster = await db.findCoasterById(id);
-
-            // If ID doesn't match any coasters, redirects to the error page
-            if (!coaster) {
-                return res.status(401).render('errors', {
-                    title: 'Error 401 - Coaster not found',
-                    message: 'Error 401 - Coaster not found in database',
-                });
-            };
-
             // If errors were found in middleware/validateCoaster.js, it will redisplay the page with the errors at the top
-            return res.status(400).render('updateCoasterForm', {
-                title: 'Update Coaster',
-                coaster: coaster,
-                errors: errors.array(),
-            });
+            return failure(res, 400, 'updateCoasterForm', `Update ${coaster.name}`, errors.array(), coaster);
         };
-
         // Receive sanitized and verified information and use it to update coaster entry in database
         const { coasterName, coasterInversions, coasterSpeed, coasterHeight, coasterLength, parkName } = matchedData(req);
         await db.updateExistingCoaster(coasterName, coasterInversions, coasterSpeed, coasterHeight, coasterLength, id, parkName);
         // Returns to main coaster page after submitting
-        res.redirect('/coaster');
-
+        return success(res, 200, '/coaster');
     } catch (err) {
+        if (err.code === '23505') {
+            return failure(res, 409, 'errors', 'Error 409 - Coaster already exists')
+        };
+
         next(err);
     };
 };
@@ -127,24 +104,14 @@ async function getSingleCoaster(req, res, next) {
     try {
         // Receives validated ID from middleware/validateId.js
         const id = req.validatedId;
-
         // Finds matching coaster in the database by ID
         const coaster = await db.findCoasterById(id);
-
         // If ID doesn't match any coasters, redirects to the error page
         if (!coaster) {
-            return res.status(401).render('errors', {
-                title: 'Error 401 - Coaster not found',
-                message: 'Error 401 - Coaster not found in database',
-            });
+            return failure(res, 401, 'errors', 'Error 401 - Coaster not found');
         };
-
         // Renders coaster data page with the selected coaster ID
-        res.render('coasterData', {
-            title: 'Coaster Data',
-            coaster: coaster,
-        });
-
+        return success(res, 200, 'coasterData', coaster.name, coaster);
     } catch (err) {
         next(err);
     };
@@ -155,12 +122,16 @@ async function deleteSingleCoaster(req, res, next) {
     try {
         // Receives validated ID from middleware/validateId.js
         const id = req.validatedId;
-
+        // Finds matching coaster in the database by ID
+        const coaster = await db.findCoasterById(id);
+        // If ID doesn't match any coasters, redirects to the error page
+        if (!coaster) {
+            return failure(res, 401, 'errors', 'Error 401 - Coaster not found');
+        };
         // Deletes matching coaster in the database by ID
         await db.deleteCoasterById(id);
         // Returns to main coaster page after deleting
-        res.redirect('/coaster');
-
+        return success(res, 204, '/coaster');
     } catch (err) {
         next(err);
     };
@@ -171,24 +142,14 @@ async function getAddToRiderForm(req, res, next) {
     try {
         // Receives validated ID from middleware/validateId.js
         const id = req.validatedId;
-
         // Finds matching coaster in the database by ID
         const coaster = await db.findCoasterById(id);
-
         // If ID doesn't match any coasters, redirects to the error page
         if (!coaster) {
-            return res.status(401).render('errors', {
-                title: 'Error 401 - Coaster not found',
-                message: 'Error 401 - Coaster not found in database',
-            });
+            return failure(res, 401, 'errors', 'Error 401 - Coaster not found');
         };
-
         // Renders add rider from coaster form
-        res.render('addRiderFromCoasterForm', {
-            title: 'Add Coaster to Rider',
-            coaster: coaster,
-        });
-
+        return success(res, 200, 'addRiderFromCoasterForm', 'Add Coaster to Rider', coaster);
     } catch (err) {
         next(err);
     };
@@ -200,26 +161,23 @@ async function postCoasterToRider(req, res, next) {
     try {
         // Receives validated ID from middleware/validateId.js
         const coasterId = req.validatedId;
-
         // Receives riderName from the post method of the add rider from coaster form
         const { riderName } = req.body;
         // Finds matching rider in the database by ID
         const riderId = await getRiderIdFromName(riderName);
-
         // If riderId doesn't match any riders, redirects to the error page
         if (!riderId) {
-            return res.status(403).render('errors', {
-                title: 'Error 403 - Rider not found',
-                message: 'Error 403 - Rider not found in database',
-            });
+            return failure(res, 401, 'errors', 'Error 401 - Rider not found')
         };
-
         // Adds link through riders_coasters join table to connect the coaster ID with the rider ID
         await addCoasterToRiderById(coasterId, riderId);
         // Redirects back to the coaster that the add rider from coaster form was initialized from
-        res.redirect(`/coaster/${coasterId}`);
-
+        return success(res, 201, `/coaster/${coasterId}`);
     } catch (err) {
+        if (err.code === '23505') {
+            return failure(res, 409, 'errors', 'Error 409 - Coaster already in rider list')
+        };
+
         next(err);
     };
 };
@@ -231,7 +189,7 @@ module.exports = {
     getSingleCoaster,
     deleteSingleCoaster,
     getUpdateCoasterForm,
-    postUpdatedCoaster,
+    putUpdatedCoaster,
     getAddToRiderForm,
     postCoasterToRider,
 };
