+++ controllers/parkController.js
@@ -1,15 +1,13 @@
 const { validationResult, matchedData } = require('express-validator');
-const db = require('../db/parkQueries');
+const db = require('../models/parkModels');
+const success = require('../utils/success');
+const failure = require('../utils/failure');
 
 // Retrieves all parks from database and displays it on parks main page
 async function getAllParks(req, res, next) {
     try {
         const parks = await db.getAllParks();
-        res.render('allParks', {
-            title: 'All Parks',
-            parks: parks,
-        });
-
+        return success(res, 200, 'allParks', 'All Parks', parks);
     } catch (err) {
         next(err);
     };
@@ -18,10 +16,7 @@ async function getAllParks(req, res, next) {
 // Retrieves the add new park form
 async function getNewParkForm(req, res, next) {
     try {
-        res.render('newParkForm', {
-            title: 'Add New Park',
-        });
-
+        return success(res, 200, 'newParkForm', 'Add New Park');
     } catch (err) {
         next(err);
     };
@@ -32,22 +27,20 @@ async function postNewPark(req, res, next) {
     try {
         // Gets results from park form validation in middleware/validatePark.js
         const errors = validationResult(req);
-
         // If there are any errors, redisplay the form page with errors at the top
         if (!errors.isEmpty()) {
-            return res.status(400).render('newParkForm', {
-                title: 'Add New Coaster',
-                errors: errors.array(),
-            });
+            return failure(res, 400, 'newParkForm', 'Add New Park', errors.array());
         };
-
         // Receive sanitized and verified information and use it to create new park entry in database
         const { parkName } = matchedData(req);
         await db.postNewPark(parkName);
         // Returns to main park page after submitting
-        res.redirect('/park');
-
+        return success(res, 201, '/park');
     } catch (err) {
+        if (err.code === '23505') {
+            return failure(res, 409, 'errors', 'Error 409 - Park already exists')
+        };
+
         next(err);
     };
 };
@@ -57,65 +50,48 @@ async function getUpdateParkForm(req, res, next) {
     try {
         // Receives validated ID from middleware/validateId.js
         const id = req.validatedId;
-
         // Finds matching park in the database by ID
         const park = await db.findParkById(id);
-
         // If ID doesn't match any parks, redirects to the error page
         if (!park) {
-            return res.status(402).render('errors', {
-                title: 'Error 402 - Park not found',
-                message: 'Error 402 - Park not found in database',
-            });
+            return failure(res, 401, 'errors', 'Error 401 - Park not found')
         };
-
         // Once park has been found, renders the update form with the park details
-        res.render('updateParkForm', {
-            title: 'Update Park',
-            park: park,
-        });
-
+        return success(res, 200, 'updateParkForm', `Update ${park.name}`, park);
     } catch (err) {
         next(err);
     };
 };
 
 // Adds updated park data to the database based on the id from the previous form retrieval
-async function postUpdatedPark(req, res, next) {
+async function putUpdatedPark(req, res, next) {
     try {
         // Gets results from park form validation in middleware/validatePark.js
         const errors = validationResult(req);
         // Receives validated ID from middleware/validateId.js
         const id = req.validatedId;
-
+        // Finds matching park in the database by ID
+        const park = await db.findParkById(id);
+        // If ID doesn't match any parks, redirects to the error page
+        if (!park) {
+            return failure(res, 401, 'errors', 'Error 401 - Park not found')
+        };
         // If there are any errors, redisplay the form page with errors at the top
         if (!errors.isEmpty()) {
-            // Finds matching park in the database by ID
-            const park = await db.findParkById(id);
-
-            // If ID doesn't match any parks, redirects to the error page
-            if (!park) {
-                return res.status(402).render('errors', {
-                    title: 'Error 402 - Park not found',
-                    message: 'Error 402 - Park not found in database',
-                });
-            };
-
             // If errors were found in middleware/validatePark.js, it will redisplay the page with the errors at the top
-            return res.status(400).render('updateParkForm', {
-                title: 'Update Park',
-                park: park,
-                errors: errors.array(),
-            });
+            return failure(res, 400, 'updateParkForm', `Update ${park.name}`, errors.array(), park);
         };
 
         // Receive sanitized and verified information and use it to update park entry in database
         const { parkName } = matchedData(req);
         await db.updateExistingPark(parkName, id);
         // Returns to main park page after submitting
-        res.redirect('/park');
-
+        return success(res, 200, '/park');
     } catch (err) {
+        if (err.code === '23505') {
+            return failure(res, 409, 'errors', 'Error 409 - Park already exists')
+        };
+
         next(err);
     };
 };
@@ -125,24 +101,14 @@ async function getSinglePark(req, res, next) {
     try {
         // Receives validated ID from middleware/validateId.js
         const id = req.validatedId;
-
         // Finds matching park in the database by ID
         const park = await db.findParkById(id);
-
         // If ID doesn't match any parks, redirects to the error page
         if (!park) {
-            return res.status(402).render('errors', {
-                title: 'Error 402 - Park not found',
-                message: 'Error 402 - Park not found in database',
-            });
+            return failure(res, 401, 'errors', 'Error 401 - Park not found');
         };
-
         // Renders park data page with the selected park ID
-        res.render('parkData', {
-            title: 'Park Data',
-            park: park,
-        });
-
+        return success(res, 200, 'parkData', park.name, park);
     } catch (err) {
         next(err);
     };
@@ -153,12 +119,16 @@ async function deleteSinglePark(req, res, next) {
     try {
         // Receives validated ID from middleware/validateId.js
         const id = req.validatedId;
-
+        // Finds matching park in the database by ID
+        const park = await db.findParkById(id);
+        // If ID doesn't match any parks, redirects to the error page
+        if (!park) {
+            return failure(res, 401, 'errors', 'Error 401 - Park not found')
+        };
         // Deletes matching park in the database by ID
         await db.deleteParkById(id);
         // Returns to the main park page after deleting
-        res.redirect('/park');
-
+        return success(res, 204, '/park');
     } catch (err) {
         next(err);
     };
@@ -169,7 +139,7 @@ module.exports = {
     getNewParkForm,
     postNewPark,
     getUpdateParkForm,
-    postUpdatedPark,
+    putUpdatedPark,
     getSinglePark,
     deleteSinglePark,
 };
